requirements.txt
Django==3.1.7
djangorestframework==3.12.4
psycopg2-binary==2.8.6
django-environ==0.4.5
minio==6.0.2
django-minio-storage==0.3.10
Pillow==8.1.2